Missed number replacements
There were two numbers which, according to the earlier hint box, should have been replaced with Xs. Additionally - and it's more of a style point - I think putting 1 and 9 for the job numbers looks cleaner than X and X+8, even if the latter is more accurate, and unlikely to cause confusion.